Reorganize toggle_led example
