+++ examples/toggle_led/interfaces/org.astarteplatform.esp32.examples.ServerDatastream.json
@@ -1,5 +1,5 @@
 {
-   "interface_name": "org.astarteplatform.esp32.ServerDatastream",
+   "interface_name": "org.astarteplatform.esp32.examples.ServerDatastream",
    "version_major": 0,
    "version_minor": 1,
    "type": "datastream",
