+++ examples/toggle_led/main/component.mk
@@ -1,1 +1,2 @@
-# Default
+COMPONENT_ADD_INCLUDEDIRS += include
+COMPONENT_SRCDIRS += src
