+++ examples/toggle_led/main/include/example_task.h
@@ -1,0 +1,11 @@
+#ifndef EXAMPLE_TASK_H
+#define EXAMPLE_TASK_H
+
+/**
+ * @brief Astarte configuration task.
+ *
+ * @param ctx ESP 32 task context.
+ */
+void astarte_example_task(void *ctx);
+
+#endif /* EXAMPLE_TASK_H */
