+++ examples/toggle_led/main/include/wifi_cfg.h
@@ -1,0 +1,9 @@
+#ifndef WIFI_CFG_H
+#define WIFI_CFG_H
+
+/**
+ * @brief Wifi module initialization function.
+ */
+void wifi_init(void);
+
+#endif /* WIFI_CFG_H */
